Library_Management_API/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-anymail==12.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
mysql-connector-python==9.0.0
packaging==24.1
python-dotenv==1.0.1
python-decouple==3.1tz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3